fix: validate noise buffer length and guard patch-embed index
GrootModel::infer and pi05 infer copy chunkSize*maxActionDim floats from
the caller-supplied noise buffer with a bare memcpy (no source-length
check), so a short Float32Array was a heap out-of-bounds read that fed
adjacent memory into the action prior. Add an exact-length check to both
the groot (patches) and pi05 (discrete) branches, mirroring the existing
images-branch guard.

Also widen the patch-embed source-index multiplication to size_t before
the multiply so the intermediate product cannot overflow int ahead of the
cast (CodeQL cpp/integer-multiplication-cast-to-long).

+++ packages/vla-ggml/addon/src/model-interface/groot.cpp
@@ -1564,7 +1564,10 @@ struct ggml_tensor* grootBuildPatchEmbedLinear(
       for (int ph = 0; ph < patch; ++ph) {
         for (int pw = 0; pw < patch; ++pw) {
           const size_t src =
-              static_cast<size_t>(pw) + patch * (ph + patch * (c + numCh * oc));
+              static_cast<size_t>(pw) +
+              static_cast<size_t>(patch) *
+                  (ph + static_cast<size_t>(patch) *
+                            (c + static_cast<size_t>(numCh) * oc));
           for (int t = 0; t < temporalPatch; ++t) {
             const int flat =
                 ((c * temporalPatch + t) * patch + ph) * patch + pw;
